search_api/requirements.txt
fastapi>=0.85.0
uvicorn[standard]>=0.20.0
elasticsearch>=7.17.0,<8.0.0
python-dotenv>=1.0.0
google-cloud-secret-manager>=2.0.0
google-cloud-bigquery>=2.0.0